+++ src/main/java/umc/puppymode/service/AuthService/KakaoService.java
@@ -1,4 +1,4 @@
-package umc.puppymode.service;
+package umc.puppymode.service.AuthService;
 
 import io.netty.handler.codec.http.HttpHeaderValues;
 import lombok.RequiredArgsConstructor;
@@ -44,11 +44,11 @@ public class KakaoService {
                 .bodyToMono(KakaoTokenResponseDTO.class)
                 .block();
 
-        log.info(" [Kakao Service] Access Token ------> {}", kakaoTokenResponseDTO.getAccessToken());
-        log.info(" [Kakao Service] Refresh Token ------> {}", kakaoTokenResponseDTO.getRefreshToken());
+//        log.info(" [Kakao Service] Access Token ------> {}", kakaoTokenResponseDTO.getAccessToken());
+//        log.info(" [Kakao Service] Refresh Token ------> {}", kakaoTokenResponseDTO.getRefreshToken());
         //제공 조건: OpenID Connect가 활성화 된 앱의 토큰 발급 요청인 경우 또는 scope에 openid를 포함한 추가 항목 동의 받기 요청을 거친 토큰 발급 요청인 경우
-        log.info(" [Kakao Service] Id Token ------> {}", kakaoTokenResponseDTO.getIdToken());
-        log.info(" [Kakao Service] Scope ------> {}", kakaoTokenResponseDTO.getScope());
+//        log.info(" [Kakao Service] Id Token ------> {}", kakaoTokenResponseDTO.getIdToken());
+//        log.info(" [Kakao Service] Scope ------> {}", kakaoTokenResponseDTO.getScope());
 
         return kakaoTokenResponseDTO.getAccessToken();
     }
@@ -70,9 +70,9 @@ public class KakaoService {
                 .bodyToMono(KakaoUserInfoResponseDTO.class)
                 .block();
 
-        log.info("[ Kakao Service ] Auth ID ---> {} ", userInfo.getId());
-        log.info("[ Kakao Service ] NickName ---> {} ", userInfo.getKakaoAccount().getProfile().getNickName());
-        log.info("[ Kakao Service ] email ---> {} ", userInfo.getKakaoAccount().getEmail());
+//        log.info("[ Kakao Service ] Auth ID ---> {} ", userInfo.getId());
+//        log.info("[ Kakao Service ] NickName ---> {} ", userInfo.getKakaoAccount().getProfile().getNickName());
+//        log.info("[ Kakao Service ] email ---> {} ", userInfo.getKakaoAccount().getEmail());
 
         return userInfo;
     }
